Isolate threading and refcount-sensitive tests into a serialized collection
Slow CI agents (Windows netfx, small Azure VMs) could starve the thread-
heavy concurrency tests when run in xUnit's parallel phase, and the exact
native-refcount singleton test could be perturbed by other tests touching
the same process-wide srgb-linear singleton concurrently.

- Extract the SKManagedStream concurrency tests into a dedicated
  SKManagedStreamConcurrencyTest in the serialized collection, and route
  their raw threads through the RunOnThread helper.
- Move SameColorSpaceCreatedDifferentWaysAreTheSameObject (exact refcount
  on the srgb-linear singleton) into SKSingletonInitTest and serialize
  that class; drain finalizers before capturing the baseline so a holder
  from the earlier parallel phase cannot decrement the count mid-test.
- Tag the remaining HandleDictionary threading classes with the
  serialized collection so nothing runs concurrently with them.

Test-only; no production code changes. Full suite: 5710 passed / 12
skipped / 0 failed.
